backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-slugify
